resgroup: generate cpu load with non-QE processes
In the test resgroup/cpu_rate_limit we need to generate cpu loads to
verify the cgroup behavior, we used to generate the cpu loads with large
sql joins, but as we now have per-QE startup memory cost in accounting,
the test might run into OOM as the join involves too many QE processes
and we simulate a small system with only 2GB ram in resource group
tests.

So we now generate the cpu loads by forking non-QE processes with python
directly.