Use Explicit Motion, instead of Redistribute, on top of Split Updates.
SplitUpdate no longer needs the old row's distribution key columns, to
calculate the segment where the old row should be deleted from. Instead,
we get the old target segment directly from the "gp_segment_id" junk
column, like a normal non-split update does. So we now use an Explicit
Motion on top Split Updates, instead of a Redistribute Motion. For the new
row, the SplitUpdate node computes the target segment from by hashing the
distribution key columns, like a Redistribute Motion does.

The reason to do this now is that with the 9.6 merge, and the
"pathification" of Split Updates, it became hard to represent the same
columns old and new value in the Motion's hash expression, before the
set_plan_references() has run. I'm sure there would be other solutions,
but this seems like a good idea anyway.

Reviewed-by: David Kimura <dkimura@pivotal.io>