Remove unnecessary dynamic cast
We were using a dynamic cast when calling Equals, but this was actually
quite expensive as we call equals very liberally. We already check that
the mdid is of type CMDIdGPDB in right above the changed line.
Additionally, we dereference the pointer right after so the dynamic cast
was useless.

Authored-by: Chris Hajas <chajas@pivotal.io>