Set gp_debug_linger to 0 for assert builds as well
gp_debug_linger is for developer convenience during interactive
testing. So, setting it to value when required is better, instead of
having it set to default for assert vs non-assert builds. As it
unnecessarily halts processes in CI as well, where no one is
monitoring or planning to attach. Any time prefer to get the hang of
process, set the gp_debug_linger and same behavior as before this
commit can be achieved, irrespective of assert or non-assert build.

This is also mainly required now more than before is for walreceiver
we are using regular error handling code same as upstream instead of
GPDB specific one. With that if any error happens in walreceiver
process, similar to upstream it would get converted to FATAL and
walreceiver will exit. But in GPDB on FATAL in assert enabled builds,
process hangs around for gp_debug_linger time (which used to be 2 mins
by default for assert builds). This causes unnecessary weird test
failures in assert builds.