Remove ExplainStateExtra compatibility hack.
The ExplainStateExtra struct was added as part of upstream commit
a5cd70dcbc, but only in back-branches. In master, changing the size of the
ExplainState struct was OK, and the hack was not needed. We picked up the
back-patched version of that commit from 9.4.1, as part of merging with the
REL9_4_STABLE, but we don't need the compatibility hack in GPDB master.
Remove it, syncing us with the version of the commit that went into
PostgreSQL master branch at the time.

Reviewed-by: Asim R P <apraveen@pivotal.io>