Rely on "-Wl,--as-needed" to filter out unused shared libs.
Starting with PostgreSQL 8.4 (upstream commit 2dad10f467), we pass
"-Wl,--as-needed", or some other platform-specific equivalent as
determined by autoconf, to the linker. It instructs the linker to not link
with shared libraries that are not used. With that, it should not be
necessary to filter them out of $(LIBS) in Makefiles.

This would be a step backwards on toolchains that don't support
"-Wl,--as-needed", but all modern toolchains do support it. We don't need
to be more portable than PostgreSQL. This reduces our diff vs upstream,
which should avoid needless merge conflicts, as we merge forward.

Reviewed-by: Paul Guo <pguo@pivotal.io>
Reviewed-by: Ning Yu <nyu@pivotal.io>