Drop some test tables at end of tests, to make regression db smaller.
The largest one was the 'dl_bt_ix' index on 'delta_blocks' table. Dropping
it saves over 600 MB. In addition to that, a bunch of other tables in the
10-50 MB range. After these changes, the regression database after running
"make installcheck-good" is about 4 GB on my laptop.

A smaller regression database is nicer to work with, if you're short on
disk space. I'm also hoping that this will make later stages of "make
installcheck-world" faster. There are some tests that take backups,
initialize standbys etc. which should go faster when there's less data.

It is generally a good idea to *not* drop test objects at end of each
test, so that they take part in the pg_upgrade testing at the end of
installcheck-world. But these test tables are not very interesting from
upgrade point of view, there is nothing special about them. We might want
to test pg_upgrade with larger relations, too, but these are not large
enough to be interesting for that either.