Remove redundant test.
There is an almost exact copy of this test at the top of 'plpgsql_cache',
just with different table and function name.

For reference, text from the JIRA ticket on this:

> If drop and re-create a (temporary) table in a PL/pgSQL function (see below for exact steps), then I get an error when calling the function a second time.
>
> I believe this is a known issue that has been fixed in PostgreSQL 8.3:
> http://wiki.postgresql.org/wiki/FAQ#Why_do_I_get_.22relation_with_OID_.23.23.23.23.23_does_not_exist.22_errors_when_accessing_temporary_tables_in_PL.2FPgSQL_functions.3F.

This was indeed fixed in PostgreSQL 8.3, with the introduction of plan
cache invalidation.

In the passing, make the formatting in qp_misc_jiras.sql a little bit
nicer.