Remove unnecessary noise from test.
No need to create a separate schema for every test table. And no need to
try dropping every table, since we're creating them in a schema; you can
easily drop the whole schema to start from clean slate, if you're running
the test locally.

Fix typo "Partiiton Table".

To be honest, these tests don't seem too useful to me. There are queries
on pg_class after each step that seem useless to me; surely we don't
expect the whole table to suddenly be gone? The test data contains
"table statistics should be kept after alter", which gives a hint on what
these test might be supposed to test, but I don't see anything to actually
verify that statistics are kept. But I'm not changing that in this commit.