Remove uninteresting 'oom_too_many_accounts' test.
The test had no comments, and the git history was unhelpful too, but I was
able to track its history in the old, archived, TINC repository. The test
was added for an old bug (MPP-25339), where a query ran out of memory. The
query in the 'oom_too_many_accounts' test was a greatly simplified version
of the original customer query, to reproduce the issue. The crux of the
issue was that we used to allocate a new Memory Accounting context in
ExecutorStart, every time a statement was run. That caused trouble, when
the main query contained a function that executed another query with SPI.
In that case, we might execute the SPI query for every row in the main
query, possibly millions of times, and created a new memory account for
every execution. Sooner or later, that lead to running out-of-memory. In
the test query, the to_timestmap() function was the SQL function, which
caused ExecutorStart to run for every row.

The original fix for that issue was to store the Memory Account in
PlannedStmt, and reuse the same memory account in subsequent calls to
ExecutorStart. But that was not the end of the story; we ran into more
similar issues later, see commit 7c9cc0539e. And later, we again added a
new explain_memory_verbosity=debug setting to actually allocate a new
memory context for every ExecutorStart call again, if that's what you want
(21f8a49108). I believe setting explain_memory_verbosity=debug would bring
back the original issue, but that's OK because it's only meant for
debugging. These later commits added more tests around this, and I believe
they cover the original case even better than the 'oom_too_many_accounts'
test. So remove 'oom_too_many_accounts' as redundant.

Discussion: https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/Tq20igUPpTQ/GCJZqmnyAgAJ
Reviewed-by: Gang Xiong <gxiong@pivotal.io>
Reviewed-by: Melanie Plageman <mplageman@pivotal.io>