Remove duplicated tests in 'DML_over_joins'.
The 'heap_motion0.sql', 'unsupported_cases0.sql' and
'partition_motion0.sql' tests are redundant with the 'heap_motion1.sql',
'unsupported_cases1.sql' and 'partition_motion1.sql' tests. The only
difference is that the 0-variants use 'int' as the datatype, and 1-variants
use 'int4'. But they are the same thing, 'int' is mapped to 'int4' early
in the parser, so we surely don't need to test both here.