Simplify cost_agg() function's signature, improve comments.
In upstream, all the callers of cost_agg that pass AGG_HASHED first
computed if the hash table would be too large to fit in work_mem. In GPDB,
the corresponding idiom is to call calcHashAggTableSizes(). Simplify the
signature to cost_agg(), so that you pass the whole HashAggTableSizes
struct, returned by calcHashAggTableSizes(), instead of extracting the
individual fields from it. That makes it easier to do the right thing and
harder to do the wrong thing; the only way to get a valid
HashAggTableSizes struct is to call calcHashAggTableSizes(), so it's not
so easy to just pass zeros anymore.

We've been carrying this difference in the cost_agg() signature as a diff
vs. upstream for a long time, through many upstream merges, but to be
honest, I never fully understood how this was supposed to work, until now.

We were previously not using this spilling-aware logic when estimating the
cost of a Hash Aggregate for Unique paths. That seemed like an oversight,
revealed by the new signature, so change it to use the spilling-aware
logic like most callers. That made the qp_dml_joins test to fail, because
we now chose a Hash Aggregate for one of the queries. Bump up
statement_mem for the test to work around that. Without setting a high
enough statement_mem, many of the test quries in the file fail in a
similar manner, this just raises the bar a little bit (see also commit
a5ff65ace5).

Improve comments in choose_hashed_setop() around the cost_agg() call.
At first, I thought that it should be using calcHashAggTableSizes() too,
and we had just forgotten to change it. But looking closer, it might be
building a Hash Setop, which doesn't spill to disk, so the upstream logic
is more appropriate. Add a comment explaining that.