Make regression test dpe unflaky.
regression/dpe case fails by accident sometimes in the
pipeline. It fails always at the same place for the following
test of plan:

  explain select * from pt, pt1 where pt.ptid = pt1.ptid and
  pt.pt1 = 'hello0' order by pt1.dist;

I found even it fails, the plan is correct, but with different
cost.

The root cause is that pt1 is distribtued randomly, then data
distribution is not stable. For some extreme cases, it may lead
to different stats info that impacts the plan cost.

This commit fixes this by creating the table distributed hashly
and after loading all the data, changing policy to randomly without
data movement. Thus every time we will have a static data distribution
plus randomly policy.