Fix DNS lookups of IPv6 addresses in URLs of web external tables.
The code to look up an IP address based on a domain name checked
'!hostname', but 'hostname' is a local array, so the condition was never
true. When compiling with -Waddress, GCC gave a warning about it:

cdbutil.c: In function ‘getDnsCachedAddress’:
cdbutil.c:1239:23: warning: the address of ‘hostinfo’ will always evaluate as ‘true’ [-Waddress]
   if (((!use_cache && !hostinfo) || (use_cache && e == NULL))
                       ^

You don't see this normally, because we set -Wno-address in configure.in,
but it sure looks wrong.

I haven't been able to find a test case where this would cause a
user-visible failure. I think you'd need a domain name that resolves to
something that's not an IPv4 address, and use that domain in an external
table's URL.