Fix cdb_sequence_nextval_qe if connection is terminated (#8941)
Fix an assert failure in cdb_sequence_nextval_qe().
When QE wants to wait a message from QD to allocate the next value,
it might receive a message with type 'X', which QD notifies QE to terminate.
Message with type 'X' is valid. So, cdb_sequence_nextval_qe should
handle it correctly.
No test for this commit. Because we need to register a fault injector in
cdb_sequence_nextval_qe() to suspend the execution of the QE process,
and later we want to resume the execution of cdb_sequence_nextval_qe()
by reset the registered fault inject. It works in normal cases. But we
need to kill the postmaster of the blocked QE process between suspending
the QE and resuming the QE. Resuming of the blocked QE failed because of
the postmaster of the QE is dead.

Co-authored-by: Paul Guo <pguo@pivotal.io>
Reviewed-by: Ning Yu <nyu@pivotal.io>