Clean up old test on pg_relation_size() function.
The test was inside a gpdiff ignore-block, so we would only notice if it
failed in a way that caused a segfault or similar. In fact, the first
test query stopped working in PostgreSQL 8.4, commit 706a308806, when the
argument type was changed from name/oid to regclass, and we hadn't
noticed. That was an intentional change, so we just need to clean up the
test.

There are plenty of pg_relation_size() tests in the 'db_size_functions'
test file already, but they all call it like "SELECT pg_relation_size()",
without any tables in the FROM clause. I think there's some value in
testing for that, like the old test queries did, so add/move a test like
that to 'db_size_functions'.