Add more information to old test query, and run it as non-superuser.
I was intrigued by this test case and the lack of comments, so I searched
for the original bug report. It seems to have been jira ticket MPP-6421.
I copied the original bug report as a comment to the test, for the benefit
of future test archeologists.

The original bug report very clearly stated that the segfault only
occurred with a non-superuser account. The original reproducer script
in the jira ticket was identical to the test we have in qp_misc_jiras.sql,
but it also had "CREATE USER" and "GRANTs", to test with a non-superuser.
I'm not sure when that aspect of the test was lost, but let's add it back
to the version we have in qp_misc_jiras.sql.

It doesn't seem likely that this exact same bug would reappear - there are
about 10 years worth of code changes between then and now - but there are
still remnants of the original fix in spi.c (search for MPP-6421 in the
comments), so I guess it's not totally out of the question. And if it's
worth testing for, let's do it faithfully to the original case.