Fix normalization of queries with window functions
Whenever there are queries involving window functions, ORCA requires the query
to be in a specific form in order to optimize it. Specifically, the query
should only contain window functions and the columns referenced by the window
clause. The algebrizer is responsible for normalizing the query to a specific
form. However the algebrizer did not handle the case when one of the columns
selected was a subquery.

For e.g

select (select b from w3 where a = w1.a) as one,
       row_number() over(partition by w1.a) as two
       from w2, w1;

In this case, the algebrizer would simply pull the subquery to the top level
without fixing the vars for the outer references.This caused the algebrizer to
crash.

This commit fixes the issue by recursing into the subquery and fixing up the
vars.