Fix gpinitsystem when using .psqlrc file
After the Postgres 9.6 merge, the .psqlrc file is always parsed if the
-X flag is not given to psql. We have a couple areas in gpinitsystem
that depend on only getting the result tuple so .psqlrc settings could
now mess with that and make gpinitsystem fail. Fix this by setting -X
flag to not read the .psqlrc file for the affected areas of
gpinitsystem.