Add the Motion to the top of the query plan earlier, as a Path.
Since PostgreSQL 9.6 and the "upper planner pathification", the planner
tends to do everything with Paths, and only generates the Plan tree at
the end of planning. In the same spirit, add the Motion to the top of
the plan, to bring the result from the QE nodes to the QD, or to take
the result to the right segments if it's INSERT/UPDATE/DELETE, before
converting the Path to Plan tree. The Path tree is generally nicer to
work with, as you have more information available in the Paths. In
particular, we still have the locus information of each Path available.

Reviewed-by: Jesse Zhang <jzhang@pivotal.io>
Reviewed-by: Pengzhou Tang <ptang@pivotal.io>
Reviewed-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>
Reviewed-by: Zhenghua Lyu <zlv@pivotal.io>