Remove hashExprs and hashOpFamilies from Flow, they are no longer needed.
These fields were used in the top-level planning, to decide if a Motion
is needed on top of the generated Plan tree. But we now do all the
top-level planning with Paths, which carry the same information as
"locus".

We still need the Flows to construct the slice table at executor startup.
It's not a very nice way to do that either, so my grand plan is to build
the slice table as part of planning, perhaps when we build the Plan tree
from the Paths in createplan.c, or perhaps as part of
set_plan_references(), or as a new step at end of planning. Not sure yet.
But for now, remove the hash expressions from the Flows, they are no longer
needed.

Reviewed-by: Jesse Zhang <jzhang@pivotal.io>
Reviewed-by: Pengzhou Tang <ptang@pivotal.io>
Reviewed-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>
Reviewed-by: Zhenghua Lyu <zlv@pivotal.io>