Fix attaching Init Plans to Motion nodes.
There were a few places in the late stages of planning and at executor
startup that assumed that a Motion node cannot have Init Plans attached to
it. And there were hacks in createplan.c, to avoid doing that. That seems
like a strange wart; there's no fundamental reason why a Motion node
couldn't or shouldn't have Init Plans. Also, the logic in createplan.c to
avoid doing that was slightly broken, and actually attached Init Plans to
multiple nodes in the plan tree, including Motion nodes. Apparently it
doesn't cause harm to have the Init Plans attached to multiple nodes in the
plan tree, because things still worked, but it sure seems wrong.

createplan.c
------------

Previously, whenever a Motion node was created, any Init Plans accumulated
so far were attached to the node below the Motion node. Revert that
change, getting rid of the create_subplan() function, and reverting
create_plan() mostly the way it is in PostgreSQL. Init Plans are now
always attached to top node of the (sub)query's plan tree.

For some reason, we were also using create_subplan() instead of
create_plan_recurse() when creating Split Update nodes. I don't know why;
perhaps a copy-pasto from the code that creates Motions. It was also
failing to save & restore curOuterRels and curOuterParams. That apparently
didn't cause any problems in practice, but seems wrong. Change it to use
create_plan_recurse() like all other create_*_plan() functions.

cdbmutate.c
-----------

Now that we attach Init Plans to Motion nodes, we have to be more careful
toalso process them in apply_motion_mutator(). Previously, we recursed
into the 'lefttree' of a Motion node directly, but skipped the targetlist,
initPlans, and other fields on a Motion node itself. That was previously
harmless because Motions didn't have Init Plans attached to them, or more
precisely, any Init Plans on Motions were *also* attached to the node
below the Motion. But now it matters. In principle, we should've scanned
the targetlist, hashExprs, quals etc. fields on a Motion node too, even
before this change, but in practice I believe that didn't cause issues
because those fields cannot contain SubPlans that are not also present in
the nodes below the Motion. But even if that doesn't cause trouble today,
let's be tidy and scan those fields, too.

execUtils.c
-----------

Executor startup makes several passes through the nodes in the plan tree,
or in the local slice of the plan tree, to find SubPlans and initplan
Params that are needed in the local slice. The walker routines to do that,
getLocallyExecutableSubplans() and ExtractParamsFromInitPlans(), had a
similar problem as apply_motion(), and didn't process a Motion node that
was a the top of the slice. The walkers don't recurse into Motions,
to restrict the scan to the local slice, but the sending Motion at the top
should be considered as part of the local slice, and now that it can
contain Init Plans, we need to scan those too.

This patch changes the EXPLAIN output of a few regression tests, that
contain Init Plans. Slice numbers have changed, because the init plans are
now encountered in different order in apply_motion().

Reviewed-by: Jesse Zhang <jzhang@pivotal.io>
Reviewed-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>