Consider Bitmap scans for OR and IN predicates on btree indexes (#554)
When we have predicates that are not AND predicates and are too
complicated for BTree index scans,
then also consider bitmap index scans. This is implemented by passing
a boolean argument, indicating that we are processing a complex predicate
and by adding an additional parameter to allow btree indexes in addition
to the primary index type, bitmap.

The change also required updating some older MDP files with the return
type of the index, since this is now required for BTree indexes as well
as for bitmap indexes (that has been the case since commit 976cba9b0).

Note that most AND predicates are handled by the btree code path ok, so there
is not need to consider them in the bitmap code path, since we could
cost the bitmap scans cheaper than the btree index scans.

There are three main cases of ANDed predicates and btree indexes:

- Multiple predicates for the same index, this is handled by the
  btree code path
- Predicates for different indexes, ORCA will pick only one index per
  alternative, so it will generate the cheapest plan that uses one index,
  while a bitmap scan would be able to use multiple indexes
- Predicates for one index plus non-index (residual) predicates, this
  is handled by the btree code path

In summary, we have an "opportunity to improve" for ANDed predicates on
multiple btree indexes. We won't fall back for such predicates and we
won't generate a bitmap scan at the moment (as planner does). This is
not a regression.