Change default for gp_statistics_use_fkeys to 'off', and deprecate it.
It's not hard to come up with an example where 'on' produces worse
estimates than 'off', but I can't come up with a realistic scenario where
'on' would produce a better plan than 'off'. Let's disable it, and save
some cycles on catalog lookups during planning. I'd like to remove it
altogether, but let's play it safe and mark it as deprecated for a release
or two first.

The documentation already claimed that the default was 'off', which was
wrong before. Move it from the "Join Operator Configuration Parameters"
section to "Other Query Planner Configuration Parameters", as it's not
really related to joins. Change the description of the GUC more accurate
in the docs, and mention that it is deprecated and will be removed in the
future.

This changed one plan in one the regression suite, in 'explain_format'
test. The query contains a join between two tables, 'apples' and 'boxes',
and the 'boxes' table is empty. (There's a third table involved in the
query, but it's not interesting for gp_statistics_use_fkeys). With
gp_statistics_use_fkeys=on, the planner estimates that the 'boxes' table
has 100000 distinct rows, because that's how many rows the 'apples' table
has. With gp_statistics_use_fkeys=off, it uses the default estimate of
1000 rows. Both esimates, of course, the table actually has 0 rows, but
it's not a very interesting case. Either plan is fine in practice.

Discussion: https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/dipq82v83PU/ro4gT1EXAgAJ
Reviewed-by: Mel Kiyama <mkiyama@pivotal.io>