Give Explicit Gather Motions new MOTIONTYPE_GATHER_SINGLE value.
An Explicit Gather Motion works differently from a normal Gather motion,
so it seems more clear to have a separate MotionType for it. It felt
wrong that the executor code was looking at the Flow information of the
subplan. I think as a matter of principle, each Plan node should carry
all the information needed to initialize the node, without peeking at
the child node.

Reviewed-by: Ning Yu <nyu@pivotal.io>