Refactor EXISTS sublink pull-up.
GPDB performs several types of optimization when pulling up EXISTS
sublink, trying to eliminate the sublink altogether if possible. This
mainly includes two cases: 1) the subquery has limit 0, 2) the subquery
has aggregates without GROUP BY or HAVING. But previously it failed to
do it correctly and would give wrong results in the case of 'limit
0/all/null' or 'offset null'.

Also previously GPDB would add LIMIT 1 to the EXISTS sublink if it is
uncorrelated and supposed to become a InitPlan, which helps gain some
performance. But this trick does not work anymore because from 9.5
(IIRC) simplify_EXISTS_query will remove the LIMIT clause if LIMIT with
a constant positive value.

This patch refactors function convert_EXISTS_sublink_to_join() to fix
the known issues as well as to make it more aligned with PostgreSQL. It
also removes the 'LIMIT 1' trick.

This patch fixes #7545

Reviewed-by: Melanie Plageman <mplageman@pivotal.io>