PXF: Do not rely on getDistributedTransactionId()
PXF relies on getDistributedTransactionId() to distribute load among
segments. This is important when the fragmentation call returns a small
list of fragments. Recently, the behavior of
getDistributedTransactionId() changed, and it no longer returns a value
for SELECT queries. As an alternative, we will use a combination of the
session id and the command count to balance the load on PXF.