packcore: find correct binary path with file command
packcore uses the file command to find the postgres binary path, it
searches for the first single-quoted string and uses it as the path.
However let's look at an output of the file command on centos 6,
manually broken into multiple lines:

    core.45685:
      ELF 64-bit LSB core file x86-64,
      version 1 (SYSV),
      SVR4-style,
      from 'postgres:  7000, gpadmin isolation2test [local] con7 cmd5 CREATE TABLE',
      real uid: 500,
      effective uid: 500,
      real gid: 501,
      effective gid: 501,
      execfn: '/usr/local/greenplum-db-devel/bin/postgres',
      platform: 'x86_64'

The first single-quoted string is actually the status line as shown in
the ps command, we should use the `execfn` field as the path.