Call TransactionIdDidCommit after XidInMVCCSnapshot only if needed
This addresses the following GPDB_96_MERGE_FIXME:
```
/*
 * GPDB_96_MERGE_FIXME: Many of the callers will call
 * TransactionIdDidCommit() after this, but if we get here, we
 * know that it committed. If we could somehow communicate it
 * to the caller, we could skip the TransactionIdDidCommit()
 * call.
 */
```
`HeapTupleSatisfiesMVCC()` frequently calls `TransactionIdDidCommit()`
right after the check: `XidInMVCCSnapshot()`. This commit tries to
reduce the number of calls necessary.

Motivation:
`HeapTupleSatisfiesMVCC()` is in a hot code path as it is the
`satisfies` callback for `Snapshot`: Also refer:
`HeapTupleSatisfiesVisibility()`. Furthermore,
`TransactionIdDidCommit()` can also be recursive if there are
subcommits.

Solution:
`XidInMVCCSnapshot()` now additionally reports if the passed in xid has
definitely committed: which we can infer from
`DISTRIBUTEDSNAPSHOT_COMMITTED_VISIBLE`, which implies that we have a
distributed committed transaction. We now check this report before
invoking `TransactionIdDidCommit()`.