Prevent DDL on temp table in read-only mode.
This exception for temp tables was added to Greenplum back in 2008.
PostgreSQL has never allowed CREATE/DROP TEMP TABLE in a read-only
transaction. The exception was made in response to a customer request
(MPP-2763):

> [Customer] is escalating this issue. When they have a system go down,
> they don't want to stop reporting activities because they are read-only
> mode. Currently they can't use Microstrategy in read-only because it uses
> temp-tables – this makes read-only mode fairly useless.
>
> Ask is to allow DDL and DML on temp tables in read-only mode so as to
> support BI activities.

There was a test script in the comments:

> gptest=# create temp table aa(a int);
> NOTICE:  Table doesn't have 'DISTRIBUTED BY' clause -- Using column named 'a'
> as the Greenplum Database data distribution key for this table.
> HINT:  The 'DISTRIBUTED BY' clause determines the distribution of data. Make
> sure column(s) chosen are the optimal data distribution key to minimize skew.
> NOTICE:  Some segments are down.  Dispatching command create temp table aa(a
> int); to remaining segments only
> CREATE TABLE
> gptest=# insert into aa values (1);
> INSERT 0 1
> gptest=# insert into aa values (2);
> INSERT 0 1
> gptest=# select * from aa;
>  a
> ---
>  1
>  2
> (2 rows)
>
> gptest=# drop table aa;
> ERROR:  Fault tolerance system has marked array read-only
> CONTEXT:  SQL statement "delete from pg_partition_rule  where paroid in  (
> select oid from pg_partition     where relid = '16499' ) or  parrelid =
> '16499' "

This test script shows some differences to how Greenplum works today.
There is no "Fault tolerance system has marked array read-only" message
anywhere in the system anymore. Back in the day, there used to be a check
in ExecCheckXactReadOnly() that threw that error:

> fail:
>    if (isFtsReadOnlySet())
>    {
>            FtsEvaluateTransientState();
>            if (isFtsReadOnlySet())
>                    ereport(ERROR,
>                                    (errcode(ERRCODE_READ_ONLY_SQL_TRANSACTION),
>                                     errmsg("Fault tolerance system has marked array read-only")));
>            else
>                    goto fail;
>    }
>    else if (XactReadOnly)
>            ereport(ERROR,
>                            (errcode(ERRCODE_READ_ONLY_SQL_TRANSACTION),
>                             errmsg("transaction is read-only")));
>    else
>            ExecutorMarkTransactionDoesWrites();

The isFtsReadOnlySet() function and the above code doesn't exist anymore.
It was removed in 2010 already, in a commit with commit message "Integrate
Release-3_4_filerep-branch -> MAIN." That commit introduced all the file
replication stuff (which was replaced with WAL replication in GPDB 6). If
I understand correctly, if the mirror goes down, we no longer consider
transactions as read-only in the master, so the original reason for
allowing CREATE/DROP on a temporary table is long gone.

Furthermore, if you try to create a temporary table in a read-only
session, there's one more hurdle that prevents it from working out of the
box:

postgres=# set default_transaction_read_only = true;
SET
postgres=# create temp table aa(a int);
ERROR:  cannot execute CREATE SCHEMA in a read-only transaction  (seg0 127.0.1.1:7002 pid=12326)
LINE 1: create temp table aa(a int);
                          ^

Even though we have the exception that allows CREATE TEMP TABLE, we still
fail to create the temporary schema. You can work around that by creating
a temporary table before making the transaction read-only, though. It
creates the temporary schema, and subsequent CREAT TEMP TABLE commands
will work.

But all in all, I think we should remove that exception, and forbid CREATE
and DROP TEMP TABLE in read-only transactions. That's what PostgreSQL
does, the original reason why we made the exception is no longer valid,
and it didn't work too well because of the temp schema issue anyway.

This patch removes the exception, and reverts the code in
check_xact_readonly() the way it is in PostgreSQL.

Reviewed-by: Asim R P <apraveen@pivotal.io>