gp_create_restore_point returns SRF with segid,lsn
Updated the catalog function gp_create_restore_point to capture the segment id
along with the restore point lsn for master and each of the segments. Returns
SETOF RECORD with fields segment_id and restore_lsn.

Note: Running the function as part of INSERT INTO SELECT * would result in a
failure due to double dispatch. The query planner has the function running on a
QE segment but the function should only be run from the QD.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>