Correctly compute distribution policy in phase 1 of gpexpand (#555)
Consider a hashed distributed partitioned table foo. After first stage
of gpexpand, the root table will have distribution polcy as hashed but
the leaf partitions will have random distribution. In such a case,
ORCA sets m_convert_hash_to_random to true in the table
descriptor. When the flag is true we should treat the distribution
policy of table as random during planning.

Note that this code will need to change when we plan to handle
gpexpand phase 1 in GPDB 6 and later. At that time, we won't be able
to treat these tables as random-partitioned anymore.

Co-authored-by: Sambitesh Dash sdash@pivotal.io
Co-authored-by: Hans Zeller hzeller@pivotal.io