Bump ORCA version to 3.85, add ICG tests for ORCA with gpexpand (#9188)
Also make minor interfance changes needed for ORCA version 3.84.

ORCA falls back at this time for tables that are in phase 1 of
gpexpand. So, right now, these tests act as a guard to ensure we
fall back. At a later point the tests might be used to validate
the correct ORCA plans.