packcore: do not remove existing pack dir
Pack dir is a temporary directory used by packcore to store the
coredump, the postgres binary and the shared libraries, then once
tarball is created the pack dir will be removed.

Packcore will error out if the pack dir already exists, the purpose is
to prevent removing the user data unexpectedly.  However even it errors
out it will still remove the pack dir during its cleanup.

Fixed the logic so an pre-existing pack dir is not changed or removed by
packcore.