Remove redundant Result nodes from ORCA-generated plans.
ORCA tends to generate redundant Result nodes, like a Result node on top
of a Result node, or a Result node on top of some other projection-capable
Result node. I'm not sure why; sometimes there are reasons for that, but in
most cases, they seem unnecessary.

If the Result only exists to project, we can easily push the projection
down to the child plan. Add a pass through the plan tree, to look for such
Results and eliminate them.

To make the mutator code work, add handler for DML nodes to
plan_tree_mutator(), and note that Sequence nodes are not capable of
projecting, in is_projection_capable_plan(). Those were existing bugs, but
they apparently didn't cause any harm before, because we never tried to
use those functions on ORCA-generated plans, and those plan nodes are only
generated by ORCA.

The reason I'm tackling this right now is that I started to work on a
patch to remove the DML node that is only used in ORCA-generated plans,
and replace it with PostgreSQL ModifyTable node. But there's one difference
between DML node and ModifyTable: DML projects the input first, and then
inserts/updates with the resulting tuple, whereas ModifyTable requires that
the input tuple is already in the correct format. One simple way to fix
that is to add a Result node under the ModifyTable to do the projection,
but that leads to even more Results. My plan is that once we have this
redundant Result elimination step, the extra Results generated in ORCA
translation would be OK, because they would usually get eliminated away
later. But even without that, having less deep plans is nicer.

Reviewed-by: Chris Hajas <chajas@pivotal.io>