Fix race condition in deadlock_under_entry_db_singleton (#9212)
Issue is immediately after backgrounded sql statement we check whether
the fault was hit. However, there is a scenario where if that statement
takes time to finish then the subsequent check to verify fault was hit
is executed too soon. Fix to use gp_wait_until_triggered_fault. This was
found in CI and is locally reproduced with following diff:

    ```
    @@ -2242,6 +2242,7 @@ StartTransaction(void)

            if (DistributedTransactionContext == DTX_CONTEXT_QE_ENTRY_DB_SINGLETON)
            {
    +               sleep(5);
                    SIMPLE_FAULT_INJECTOR("transaction_start_under_entry_db_singleton");
            }
    ```