Delay fsync on appendoptimized tables until next restartpoint
Without this patch, for every appendoptimized insert WAL record that
is replayed, an fsync system call is made.  As known that fsync is a
costly system call which could block the replay process and make the
replay process much slower than expected.

This patch changes it so that the replay process defers fsync requests
to the checkpointer process on mirror.  The replay process only
performs write.  Note that the behavior on primary is not changed.  An
appendoptimized segment file is fsync'ed by a primary backend at the
end of each insert.

Co-authored-by: Paul Guo <paulguo@gmail.com>