Harden analyze rootpartition to merge stats if relpages is incorrect (#9213)
`analyze rootpartition` will merge statistics if all leaf partitions
have been analyzed. If a partition is empty (that is, the reltuples is
0), we previously considered it analyzed only if the relpages was 1.

However, we have seen a case where the reltuples is 0 and relpages was
greater than 1 (one such case was resolved in 183947f). This commit hardens
analyze in case we encounter such a scenario, as it can make `analyze rootpartition`
take significantly longer.