Fix gp_statistics_pullup_from_child_partition.
The code to fetch the largest partition's statistics was broken in the
9.6 merge. Fix it.

There are no regression tests for
gp_statistics_pullup_from_child_partition=on, but the codepath where the
parent has no statistics is exercised by existing tests. I guess that's
enough for now. (I still feel that we should probably get rid of this
altogether, as I mentioned in the old comment, but that's a different
story.)