Rewrite test case xlog files (#9192)
* Revert "Increase default value of wal_keep_segments GUC."
This reverts commit dd18c4a0497b51f096708e0f867c2734391be348.
We have replication slot now, wal_keep_segments is not needed for
safely removing WAL files.

* Rewrite test missing_xlog
Old missing_xlog tries to test an error happened on the primary
when requesting for WAL files from the mirror are unavailable.
To simulate this scenario, the test moves WAL files to other folder,
encounter the error and moves them back. But it may damage the WAL
files on the primary, because it assumes that there is no WAL write
between the 2 movings of WAL files. Unluckily, the assumption is false.
The hint bits could trigger a WAL write request(FPI_FOR_HINT), and
the hot standby may also requst a WAL write. So moving WAL files is
not recommended.

To make `sync_error` occur, we set `wal_keep_segments` to 0,
i.e. disable this feature. And in order to temporarily break
the constraint of replication slot, to delete WAL files used by
the mirror, we use fault inject. In the end, the mirror loses some
WAL file to do incremently recovery, we do full recovery for the mirror.

* Add 'wal_keep_segments = 5' to configuration of test pg_rewind
Since wal_keep_segments is set to 0 by default, WAL files are
only guarded by replication slot. In test case pg_rewind, there is
no replication slot found now. A workaround way is to set
wal_keep_segments back to 5 in the test case.


Co-authored-by: Asim R P <apraveen@pivotal.io>