Don't set bogus flow on SplitUpdate nodes.
The 'numsegments' in a SplitUpdate's flow was used in very a bizarre way.
It didn't indicate the number of segments used to execute the SplitUpdate
node itself, instead, it indicated the number of segments in the target
table. In createplan.c, that number was then propagated up to the Explicit
Redistribute Motion node above it, via the flows. Set the number of target
segments explicitly in the SplitUpdate node, and stop abusing the Flow for
that. The Flow is now left uninitialized on the SplitUpdate node, which
means that it will be propagated from the child in apply_motion(), like
with most plan nodes. And the Motion gets the number of segments from its
path's locus, like most plan nodes.