Rely on junk status to find 'action' and 'tupleoid' columns in ModifyTable.
We've always relied on the junk column's name to find the 'ctid' column,
like in PostgreSQL. And also for the GPDB-specific 'gp_segment_id' column.
But for some reason, the 'action' and 'oid' columns used a different
mechanism, where the column indexes were stored in the ModifyTable plan
node. With those two, we furthermore had two slightly different ways to
find them at planning time. For the 'oid' column, we relied on the junk
column name, like for the 'ctid' column in upstream, except that it was
extracted at plan time instead of executor startup time. For the 'action'
column, the planner relied on the fact that there should only be one
DMLActionExpr in the target list. I don't know why we didn't rely on the
column name for that too.

Clean that up by consistently using the resjunk colum names to find all
of these columns, like we do for 'ctid' in PostgreSQL.

Reviewed-by: Hans Zeller <hzeller@pivotal.io>