Merge DML node into ModifyTable.
Historically, ORCA has used a DML node to perform inserts/updates/deletes
on a table. PostgreSQL 9.0 introduced a similar node type, ModifyTable,
but we didn't get around to refactor ORCA to use the new upstream node
type. The DML node did share the low-level code (ExecInsert/ExecDelete)
to modify the table, but it was still a separate node type. This patch
removes the DML node, and refactors the ORCA translator code to translate
ORCA physical DML nodes to ModifyTable nodes instead.

There are some small differences between DML and ModifyTable nodes.
ModifyTable has extra capabilities to do INSERT ON CONFLICT, for example,
that ORCA doesn't need because it doesn't support INSERT ON CONFLICT. The
ModifyTable node also behaves slightly differently, when it's used in ORCA
plans: it computes the target partition for a DELETE based on the key
values, whereas the postgres planner has a completely different strategy
for UPDATEs and DELETEs on a partitioned table (it generates a separate
sub-plan for each partition). But that difference in behavior was already
in the ExecDelete code, which was already shared between DML and
ModifyTable nodes.

The major difference between DML and ModifyTable nodes that this patch
needs to work around is that the DML node was able to project, before
inserting a tuple, while ModifyTable requires that the input tuple coming
from the node below has the correct physical format for the table,
including the presence of dropped columns. To work around that difference,
the ORCA translator now adds a Result if necessary to do the projection.
Or to be precise, it always adds the Result node, but it gets stripped off
in the redundant Result removal phase later, if it's not needed.

Another minor difference is that ModifyTable relies on junk column names
to find the 'ctid', 'gp_segment_id', 'DMLAction', and 'oid' columns, while
DML node stored the indexes of those columns in the DML plan node. Teach
the ORCA translator to label those columns as junk columns with the names
that the ModifyTable executor code expects.

Reviewed-by: Hans Zeller <hzeller@pivotal.io>