Use correct accessor methods
CStringGetDatum operates on const char pointers, but restore_name is
a TEXT pointer. Use the PointerGetDatum accessor instead. restore_ptr
is an XLogRecPtr, so the correct accessor there is LSNGetDatum. This
fixes a compiler warning on passing incompatible pointer types.

Also apply the DatumGetLSN accessor on the output of the function call
as it's stored in an XLogRecPtr and fix a copy/pasto in the file header
identification.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>
Reviewed-by: Jimmy Yih <jyih@pivotal.io>